fix: allow disabling tracking of recent buffers
Previously, setting the `buflist` config option to an empty list would
result in an OOB index when calling `shift_buflist`. This is now fixed
by checking the length beforehand resp. disabling the calling
autocommand entirely.

As disabling the tracking of buffers results in an initially empty
mark-list (and thus tabline) after loading the plugin, an additional
config option `always_show_tabline` is now available, which can be
toggled to defer tabline initialization until *after* a buffer is
marked.

Closes: https://github.com/iofq/dart.nvim/issues/1